+++ packages/history/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @xnetjs/history
 
+## 2.1.0
+
+### Patch Changes
+
+- Updated dependencies [[`0a4a1de`](https://github.com/crs48/xNet/commit/0a4a1de41b0f68c197ba5f7d191706668550f708)]:
+  - @xnetjs/data@2.1.0
+  - @xnetjs/sync@2.1.0
+  - @xnetjs/core@2.1.0
+
 ## 2.0.0
 
 ### Minor Changes